вынес функции для работы с массивами строк в array_utils.c
